+++ platform/common/utils/hash/hasher.go
@@ -6,12 +6,6 @@ SPDX-License-Identifier: Apache-2.0
 
 package hash
 
-import (
-	"hash"
-)
-
 type Hasher interface {
-	GetHash() hash.Hash
-
 	Hash(msg []byte) ([]byte, error)
 }
